remove Provisioner.stop_defer()
This was causing headaches for non-trivial cleanups, ie. if it was
to be implemented for LibvirtCloningProvisioner, it would need to
break the existing connection (to stop background threads cloning
VM images), then re-connect, parallelize freeing VMs, then join
back together before closing the shared connection.

This isn't possible with a flat list of callables given by the
.stop_defer() API function.

As such, leave any parallel optimizations to the Provisioner itself
and its own .stop().

Signed-off-by: Jiri Jaburek <comps@nomail.dom>